pyproject.toml
[tool.poetry]
name = "dynamic-dns-digital-ocean"
version = "0.1.0"
description = "Dyanmic DNS updates for Digital Ocean"
authors = ["Zane Williamson <zane@brodate.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23"
python-digitalocean = "^1.15"
click = "^7.1"
python-daemon = "^2.2"
tldextract = "^2.2"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
ipython = "^7.14"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
